Implementato blocco nella cancellazione di utenti admin

+++ app/controllers/settings/users_controller.rb
@@ -3,6 +3,9 @@ class Settings::UsersController < Settings::BaseController
   end
 
   def destroy
+    if Current.user.admin?
+      redirect_to settings_user_path, alert: "Admin users cannot delete their account." and return
+    end
     terminate_session
     Current.user.destroy
     redirect_to root_path, notice: "Your account has been deleted."
